+++ servlet/src/main/java/com/example/filter/CharacterEncodingFilter.java
@@ -1,9 +1,10 @@
-package com.example;
+package com.example.filter;
 
 import jakarta.servlet.*;
 import jakarta.servlet.annotation.WebFilter;
 
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 
 @WebFilter("/*")
 public class CharacterEncodingFilter implements Filter {
@@ -11,6 +12,7 @@ public class CharacterEncodingFilter implements Filter {
     @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
         request.getServletContext().log("doFilter() 호출");
+        response.setCharacterEncoding(StandardCharsets.UTF_8.displayName());
         chain.doFilter(request, response);
     }
 }
