+++ servlet/src/main/java/com/example/servlet/SharedCounterServlet.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.servlet;
 
 import jakarta.servlet.ServletConfig;
 import jakarta.servlet.ServletException;
@@ -8,6 +8,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Servlet 인터페이스의 service 메서드는 HTTP 요청 및 응답을 처리할 때 사용된다.
@@ -30,20 +31,20 @@ public class SharedCounterServlet extends HttpServlet {
      * 서버는 여러 스레드에서 접근 가능하므로 서블릿에서 비즈니스 로직을 처리할 때 인스턴스 변수는 사용하지 않는다.
      * 다른 사용자에게 공유되어도 문제가 없는 불변 객체라면 서블릿의 인스턴스 변수로 사용 가능하다.
      */
-    private Integer sharedCounter;
+    private AtomicInteger sharedCounter;
 
     @Override
     public void init(final ServletConfig config) throws ServletException {
         super.init(config);
         getServletContext().log("init() 호출");
-        sharedCounter = 0;
+        sharedCounter = new AtomicInteger(0);
     }
 
     @Override
     protected void service(final HttpServletRequest request, final HttpServletResponse response) throws IOException {
         getServletContext().log("service() 호출");
-        sharedCounter++;
-        response.getWriter().write(String.valueOf(sharedCounter));
+        final int result = sharedCounter.incrementAndGet();
+        response.getWriter().write(String.valueOf(result));
     }
 
     @Override
